requirements.txt
-e git+git@github.com:jarlaxle-536/evensound.git@f96a85ddf9529df8916366d74b34290a39c6c9bf#egg=evensound
Faker==4.14.2
mido==1.2.9
Pillow==8.0.1
pk==1.0
pkg-resources==0.0.0
py-midi==2.0.1
pygame==2.0.0
pyserial==3.4
python-dateutil==2.8.1
python-rtmidi==1.4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
text-unidecode==1.3
vext==0.7.4
vext.pyqt5==0.7.4
